kitchen_agent/requirements.txt
langchain-google-genai>=0.0.5
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
chromadb>=0.4.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.0.0
